submit java-oop-ru/interfaces

+++ java-oop-ru/interfaces/src/main/java/exercise/Home.java
@@ -3,6 +3,6 @@ package exercise;
 // BEGIN
 public interface Home {
     double getArea();
-    double compareTo(Home home);
+    int compareTo(Home home);
 }
 // END
